Move TreeTracker utilities to their own file
